requirements.txt
ffmpeg-python==0.2.0
absl-py==1.3.0
aiosignal==1.2.0
alembic==1.8.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asteroid-filterbanks==0.4.0
async-timeout==3.0.1
attrs==22.1.0
audioread==3.0.0
autopage==0.5.1
backports.cached-property==1.0.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cliff==4.0.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.5
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
distlib==0.3.5
docopt==0.6.2
einops==0.3.2
fastapi==0.86.0
ffmpeg==1.4
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
greenlet==2.0.0
grpcio==1.50.0
gunicorn==20.1.0
h11==0.14.0
hmmlearn==0.2.8
huggingface-hub==0.10.1
HyperPyYAML==1.0.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mpmath==1.2.1
multidict==4.7.6
networkx==2.8.8
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.2.3
optuna==3.0.3
packaging==21.3
pandas==1.5.1
pbr==5.11.0
Pillow==9.3.0
platformdirs==2.5.2
pooch==1.6.0
prettytable==3.5.0
primePy==1.3
protobuf==3.19.6
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
#python==3.11.1
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytorch-metric-learning==1.6.3
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Rx==3.2.0
scikit-learn==1.1.3
scipy==1.8.1
semver==2.13.0
sentencepiece==0.1.97
shellingham==1.5.0
simplejson==3.17.6
singledispatchmethod==1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.5
SoundFile==0.10.3.post1
speechbrain==0.5.13
SQLAlchemy==1.4.42
srt==3.5.2
sse-starlette==1.1.6
starlette==0.20.4
stevedore==4.1.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.13.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tqdm==4.64.0
typer==0.6.1
typing_extensions==4.4.0
urllib3==1.26.11
uvicorn==0.19.0
virtualenv==20.16.3
vosk==0.3.42
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
yarl==1.5.1
zipp==3.8.1
pyhassbian==0.2.1
aiohttp==3.6.3
aiohttp-basicauth==1.0.0
aiohttp-cors==0.6.0
python-socketio==4.6.0
python-engineio==3.13.2

